NodeInfo page
Made node info page passing the id of the selected node. I made a route to the page. I have added an id tag to my array in home.js that fakes the API in a simple way.

(contains the same array that fakes the API in a simple way)